Bugfix: the end of a reorged chain is invalid when connect fails
Introduced in 19c182ce0322c0c56c35d560d981833ad6105e9e

When an invalid block is found during a reorg, we know the last of the blocks in
the was-to-be-connected chain is invalid, but not necessarily the first. As
vpIndexToConnect is ordered in decreasing height, the end of the reorg is the
front of the vector, and not the back.

This only affected the warning system.