[addrman] Fix new table bucketing during unserialization
An addrman entry can appear in up to 8 new table buckets. We store this
entry->bucket indexing during shutdown so that on restart we can restore
the entries to their correct buckets.

Commit a692bd8b01e44b732052f0a87369ecb8f9bc79e4 broke the
deserialization code so that each entry could only be put in up to one
new bucket. Fix that.